feat: 검색 키워드 강조 함수 위치 수정 - #162, #163, #168
