+++ frontend/components/search/ArticleItem/index.tsx
@@ -14,41 +14,17 @@ import {
 } from './styled';
 
 interface ArticleItemProps {
-  keyword: string;
   title: string;
   content: string;
   nickname: string;
 }
 
-export default function ArticleItem({ keyword, title, content, nickname }: ArticleItemProps) {
-  const highlightWord = (text: string, words: string[]): React.ReactNode => {
-    let wordIndexList = words.map((word) => text.toLowerCase().indexOf(word.toLowerCase()));
-
-    const filteredWords = words.filter((_, index) => wordIndexList[index] !== -1);
-    wordIndexList = wordIndexList.filter((wordIndex) => wordIndex !== -1);
-
-    if (wordIndexList.length === 0) return text;
-
-    const startIndex = Math.min(...wordIndexList);
-
-    const targetWord = filteredWords[wordIndexList.indexOf(startIndex)];
-
-    const endIndex = startIndex + targetWord.length;
-
-    return (
-      <>
-        {text.slice(0, startIndex)}
-        <b>{text.slice(startIndex, endIndex)}</b>
-        {highlightWord(text.slice(endIndex), words)}
-      </>
-    );
-  };
-
+export default function ArticleItem({ title, content, nickname }: ArticleItemProps) {
   return (
     <ItemWrapper>
       <ItemGroup>
-        <ItemTitle>{highlightWord(title, keyword.trim().split(' '))}</ItemTitle>
-        <ItemContent>{highlightWord(content, keyword.trim().split(' '))}</ItemContent>
+        <ItemTitle>{title}</ItemTitle>
+        <ItemContent>{content}</ItemContent>
       </ItemGroup>
       <UserProfile>
         <ProfileDescription>
